Adding a Class that will run the blockchain on a client.
